fix: harden kata boundaries and validation (#61)
* fix: harden kata boundaries and validation

* ci: prime surefire before offline gate

* ci: quote Maven selector on Windows

+++ src/main/java/kyu6/TwoSum.java
@@ -8,7 +8,7 @@ public class TwoSum {
     public static int[] twoSum(int[] numbers, int target) {
         for (int i = 0; i < numbers.length; i++) {
             for (int j = i + 1; j < numbers.length; j++) {
-                if (numbers[i] + numbers[j] == target) {
+                if ((long) numbers[i] + numbers[j] == target) {
                     return new int[]{i, j};
                 }
             }
